Put interaction components and utils in a dedicated folder
