+++ apps/storybook/src/YZoom.stories.tsx
@@ -1,5 +1,5 @@
 import { Pan, ResetZoomButton, VisCanvas, YAxisZoom } from '@h5web/lib';
-import type { ModifierKey as ModifierKeyType } from '@h5web/lib/src/vis/models';
+import type { ModifierKey as ModifierKeyType } from '@h5web/lib';
 import type { Meta, Story } from '@storybook/react';
 
 import VisCanvasStoriesConfig from './VisCanvas.stories';
