+++ packages/lib/src/interactions/ResetZoomButton.module.css
@@ -6,14 +6,14 @@
 }
 
 .btn {
-  composes: btn from '../../toolbar/Toolbar.module.css';
+  composes: btn from '../toolbar/Toolbar.module.css';
   padding: 0.5rem !important; /* FIX style ordering issue with Vite */
   pointer-events: auto;
   font-size: 0.875em;
 }
 
 .btnLike {
-  composes: btnLike from '../../toolbar/Toolbar.module.css';
+  composes: btnLike from '../toolbar/Toolbar.module.css';
   background-color: rgba(255, 255, 255, 0.7);
   box-shadow: rgba(0, 0, 0, 0.1) 0px 0px 0px 1px,
     rgba(0, 0, 0, 0.1) 0px 4px 11px;
