+++ packages/lib/src/toolbar/Toolbar.tsx
@@ -3,7 +3,7 @@ import { useMap, useMeasure } from '@react-hookz/web';
 import type { ReactElement, ReactNode } from 'react';
 import flattenChildren from 'react-keyed-flatten-children';
 
-import type { Interaction } from '../vis/models';
+import type { Interaction } from '../interactions/models';
 import MeasuredControl from './MeasuredControl';
 import OverflowMenu from './OverflowMenu';
 import Separator from './Separator';
