+++ packages/lib/src/vis/models.ts
@@ -8,7 +8,6 @@ import type {
   NumArray,
 } from '@h5web/shared';
 import type { PickD3Scale, PickScaleConfigWithoutType } from '@visx/scale';
-import type { Vector2, Vector3 } from 'three';
 
 import type { ColorMap } from './heatmap/models';
 import type { ScaleGamma } from './scaleGamma';
@@ -91,30 +90,6 @@ export interface HistogramParams {
   bins: NumArray;
   colorMap?: ColorMap;
   invertColorMap?: boolean;
-}
-
-export type ModifierKey = 'Alt' | 'Control' | 'Shift';
-
-export interface Selection {
-  startPoint: Vector2;
-  endPoint: Vector2;
-}
-
-export interface CanvasEvent<T extends PointerEvent | WheelEvent> {
-  unprojectedPoint: Vector3;
-  sourceEvent: T;
-}
-
-export interface CanvasEventCallbacks {
-  onPointerDown?: (evt: CanvasEvent<PointerEvent>) => void;
-  onPointerMove?: (evt: CanvasEvent<PointerEvent>) => void;
-  onPointerUp?: (evt: CanvasEvent<PointerEvent>) => void;
-  onWheel?: (evt: CanvasEvent<WheelEvent>) => void;
-}
-
-export interface Interaction {
-  shortcut: string;
-  description: string;
 }
 
 export interface CustomColor {
